+++ test/android_id_test.dart
@@ -6,13 +6,14 @@ import 'package:flutter_test/flutter_test.dart';
 void main() {
   const plugin = AndroidId();
   const channel = MethodChannel('android_id');
+  const mockAndroidId = '42';
 
   TestWidgetsFlutterBinding.ensureInitialized();
   debugDefaultTargetPlatformOverride = TargetPlatform.android;
 
   setUp(() {
     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-        .setMockMethodCallHandler(channel, (methodCall) async => '42');
+        .setMockMethodCallHandler(channel, (methodCall) async => mockAndroidId);
   });
 
   tearDown(() {
@@ -21,6 +22,34 @@ void main() {
   });
 
   test('getAndroidId', () async {
-    expect(await plugin.getId(), '42');
+    expect(await plugin.getId(), mockAndroidId);
+  });
+
+  test('returns null on non-Android platforms', () async {
+    addTearDown(() => debugDefaultTargetPlatformOverride = null);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.fuchsia;
+
+    expect(await plugin.getId(), isNull);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+
+    expect(await plugin.getId(), isNull);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.linux;
+
+    expect(await plugin.getId(), isNull);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.macOS;
+
+    expect(await plugin.getId(), isNull);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.windows;
+
+    expect(await plugin.getId(), isNull);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.android;
+
+    expect(await plugin.getId(), mockAndroidId);
   });
 }